Reworking src/ dir to be a little more clean generate and parse are big enough that I've added CMakeLists for each They output .a files
liblibgear.so is a library that will eventually be used for a possible LSP
gearlang->gear
