+++ .gitignore
@@ -4,14 +4,16 @@ Makefile
 cmake_install.cmake
 CMakeCache.txt
 .cache/
+/parse/
+/generate/
 # And ninja
 .ninja_deps
 .ninja_log
 build.ninja
 
 # The executables
-/gearlang
-/gearlang.exe
+/gear
+/liblibgear.so
 
 # Generated output
 build/
