+++ src/generate/var.cpp
@@ -30,45 +30,68 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE  OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#pragma once
-
-#include <memory>
-#include <cassert>
-
 #include <llvm/IR/Value.h>
+#include <llvm/IR/Constants.h>
+#include <llvm/IR/GlobalVariable.h>
+#include <llvm/IR/Type.h>
 
+#include <gearlang/ast/expr.hpp>
+#include <gearlang/ast/stmt.hpp>
+#include <gearlang/ast/func.hpp>
+#include <gearlang/ast/vars.hpp>
 
-/// @brief Safe casting
-template<typename From, typename To>
-inline To* try_cast(From* from) {
-    assert(from != nullptr);
-    return dynamic_cast<To*>(from);
+#include <gearlang/error.hpp>
+#include <gearlang/etc.hpp>
+
+using namespace Ast::Nodes;
+
+inline llvm::Value* deref(
+    Context& ctx, 
+    llvm::Value* ptr, 
+    Sem::Type* type,
+    const std::string& name,
+    const char* suffix
+) {
+    return ctx.builder.CreateLoad(
+        type->to_llvm(ctx),
+        ptr,
+        name + suffix
+    );
 }
 
-/// @brief Wrapper for try_cast to specifically take input from a unique_ptr*
-template<typename From, typename To>
-inline To* cast_from_uptr(std::unique_ptr<From>* from) {
-    From* underlying = from->get();
-
-    return try_cast<From, To>(underlying);
+llvm::Value* get_var(NodeBase* let) {
+    if(Let* let_val = try_cast<NodeBase, Let>(let))
+        return let_val->var;
+    
+    return try_cast<NodeBase, Argument>(let)->var;
 }
 
-/// @brief span metadata for tokens
-struct Span {
-    size_t line;
-    size_t col;
-    size_t start;
-    size_t end;
-};
 
-struct Options {
-    std::string input;
-    std::string output;
+// Looks up the name of the variable
+// If the variable doesn't exist, it throws an error and quits
+// Otherwise Checks if it is a global and returns it
+llvm::Value* ExprVar::generate(Context& ctx) {
+    return deref(
+        ctx, access_alloca(ctx),
+        ty.get(), name, ".load"
+    );
+}
 
-    bool verbose = false;
-    bool emit_object = false;
-    bool emit_llvm = false;
+llvm::Value* ExprVar::access_alloca(Context& _ctx) {
+    return get_var(let);
+}
 
-    bool dump_tokens = false;
-    bool dump_ast = false;
-};
+llvm::Value* ExprStructParam::generate(Context& ctx) {
+    llvm::Type* ty_ll = ty->struct_param_ty(index).to_llvm(ctx);
+
+    return ctx.builder.CreateLoad(ty_ll, access_alloca(ctx), struct_name + "." + name + ".load");
+}
+
+// Returns the address calculated from the GEP
+llvm::Value* ExprStructParam::access_alloca(Context& ctx) {
+    llvm::Type* struct_ll = ty->get_llvm_struct();
+
+    return ctx.builder.CreateStructGEP(
+        struct_ll, get_var(let), index+1
+    );
+}
